- fixing scope issue, still not clear how this stopped working
