+++ Ollivanders/src/net/pottercraft/ollivanders2/spell/O2Spells.java
@@ -63,68 +63,7 @@ public class O2Spells
     /**
      * Spell allow/disallow zones.
      */
-    final private ArrayList<SpellZone> spellZones = new ArrayList<>();
-
-    /**
-     * String keys for zone config
-     */
-    private final static String allowedList = "allowed-spells";
-    private final static String disallowList = "disallowed-spells";
-    private final static String globalZoneName = "global";
-    private final static String none = "NONE";
-    private final static String all = "ALL";
-
-    /**
-     * Type of spell zones that can be defined
-     */
-    public enum SpellZoneType
-    {
-        CUBOID,
-        WORLD,
-        WORLD_GUARD;
-    }
-
-    /**
-     * A zone that spells can be explicitly allowed and disallowed for
-     */
-    public class SpellZone
-    {
-        String zoneName;
-        SpellZoneType zoneType;
-        final Cuboid cuboid;
-        String zoneWorldName;
-
-        ArrayList<O2SpellType> disallowedSpells;
-        ArrayList<O2SpellType> allowedSpells;
-
-        /**
-         * Constructor
-         *
-         * @param name the name of this zone
-         * @param world the name of the world this zone is in
-         * @param type the type of zone
-         * @param area the area bounds for this zone if a cuboid, this should be opposite corners
-         * @param allowed a list of allowed spells, if set, this will take precedence over a disallow list
-         * @param disallowed a list of disallowed spells
-         */
-        SpellZone(@NotNull String name, @NotNull String world, @NotNull O2Spells.SpellZoneType type, int[] area, @NotNull ArrayList<O2SpellType> allowed, @NotNull ArrayList<O2SpellType> disallowed)
-        {
-            zoneName = name;
-            zoneType = type;
-            zoneWorldName = world;
-
-            if (type == SpellZoneType.CUBOID)
-                cuboid = new Cuboid(zoneWorldName, area);
-            else
-            {
-                int[] emptyArea = {0, 0, 0, 0, 0, 0};
-                cuboid = new Cuboid(zoneWorldName, emptyArea);
-            }
-
-            disallowedSpells = disallowed;
-            allowedSpells = allowed;
-        }
-    }
+    final ArrayList<SpellZone> spellZones = new ArrayList<>();
 
     /**
      * Constructor
@@ -232,12 +171,12 @@ public class O2Spells
 
         for (String zone : zoneConfig.getKeys(false))
         {
-            if (zone.equalsIgnoreCase(globalZoneName))
+            if (zone.equalsIgnoreCase(SpellZone.globalZoneName))
             {
                 common.printDebugMessage("Loading global zone config:", null, null, false);
 
-                globalAllowedSpells = getSpellsForZone(zone, allowedList);
-                globalDisallowedSpells = getSpellsForZone(zone, disallowList);
+                globalAllowedSpells = getSpellsForZone(zone, SpellZone.allowedList);
+                globalDisallowedSpells = getSpellsForZone(zone, SpellZone.disallowList);
             }
             else
             {
@@ -260,10 +199,10 @@ public class O2Spells
         if (typeString == null || typeString.length() < 1)
             return;
 
-        SpellZoneType type;
+        SpellZone.SpellZoneType type;
         try
         {
-            type = SpellZoneType.valueOf(typeString.toUpperCase());
+            type = SpellZone.SpellZoneType.valueOf(typeString.toUpperCase());
         }
         catch (Exception e)
         {
@@ -272,7 +211,7 @@ public class O2Spells
         }
 
         String world = "";
-        if (type == SpellZoneType.WORLD || type == SpellZoneType.CUBOID)
+        if (type == SpellZone.SpellZoneType.WORLD || type == SpellZone.SpellZoneType.CUBOID)
         {
             world = zoneConfig.getString(zoneName + "." + "world");
 
@@ -285,7 +224,7 @@ public class O2Spells
 
         int[] area = {0, 0, 0, 0, 0, 0};
 
-        if (type == SpellZoneType.CUBOID)
+        if (type == SpellZone.SpellZoneType.CUBOID)
         {
             String areaString = zoneConfig.getString(zoneName + "." + "area");
 
@@ -303,8 +242,8 @@ public class O2Spells
             }
         }
 
-        ArrayList<O2SpellType> allowed = getSpellsForZone(zoneName, allowedList);
-        ArrayList<O2SpellType> disallowed = getSpellsForZone(zoneName, disallowList);
+        ArrayList<O2SpellType> allowed = getSpellsForZone(zoneName, SpellZone.allowedList);
+        ArrayList<O2SpellType> disallowed = getSpellsForZone(zoneName, SpellZone.disallowList);
 
         SpellZone zone = new SpellZone(zoneName, world, type, area, allowed, disallowed);
         spellZones.add(zone);
@@ -371,7 +310,7 @@ public class O2Spells
         // check world permissions
         for (SpellZone zone : spellZones)
         {
-            if (zone.zoneType != SpellZoneType.WORLD)
+            if (zone.zoneType != SpellZone.SpellZoneType.WORLD)
                 continue;
 
             World world = location.getWorld();
@@ -397,7 +336,7 @@ public class O2Spells
         // check world guard zone permissions
         for (SpellZone zone : spellZones)
         {
-            if (zone.zoneType != SpellZoneType.WORLD_GUARD)
+            if (zone.zoneType != SpellZone.SpellZoneType.WORLD_GUARD)
                 continue;
 
             if (Ollivanders2.worldGuardO2.isLocationInRegionByName(zone.zoneName, location) && zone.allowedSpells.size() > 0)
@@ -416,7 +355,7 @@ public class O2Spells
         // check cuboid zone permissions
         for (SpellZone zone : spellZones)
         {
-            if (zone.zoneType != SpellZoneType.CUBOID)
+            if (zone.zoneType != SpellZone.SpellZoneType.CUBOID)
                 continue;
 
             if (zone.cuboid.isInside(location, common) && zone.allowedSpells.size() > 0)
@@ -458,7 +397,7 @@ public class O2Spells
         // check world permissions
         for (SpellZone zone : spellZones)
         {
-            if (zone.zoneType != SpellZoneType.WORLD)
+            if (zone.zoneType != SpellZone.SpellZoneType.WORLD)
                 continue;
 
             World world = location.getWorld();
@@ -477,7 +416,7 @@ public class O2Spells
         // check world guard zone permissions
         for (SpellZone zone : spellZones)
         {
-            if (zone.zoneType != SpellZoneType.WORLD_GUARD)
+            if (zone.zoneType != SpellZone.SpellZoneType.WORLD_GUARD)
                 continue;
 
             if (Ollivanders2.worldGuardO2.isLocationInRegionByName(zone.zoneName, location) && zone.disallowedSpells.contains(spellType))
@@ -489,7 +428,7 @@ public class O2Spells
         // check cuboid zone permissions
         for (SpellZone zone : spellZones)
         {
-            if (zone.zoneType != SpellZoneType.CUBOID)
+            if (zone.zoneType != SpellZone.SpellZoneType.CUBOID)
                 continue;
 
             if (zone.cuboid.isInside(location, common) && zone.disallowedSpells.contains(spellType))
